Юнит тесты, форматирование во всех файлах.

+++ VerstaTask/Models/OrderEditDto.cs
@@ -6,26 +6,26 @@
 
     public class OrderEditDto
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
 
         [Required(ErrorMessage = "Поле обязательно")]
         [DisplayName("Город отправителя")]
-        [DataType(DataType.Text,ErrorMessage = "Неправильные данные!")]
+        [DataType(DataType.Text, ErrorMessage = "Неправильные данные!")]
         public string SenderCity { get; set; }
 
         [Required(ErrorMessage = "Поле обязательно")]
         [DisplayName("Адрес отправителя")]
-        [DataType(DataType.Text,ErrorMessage = "Неправильные данные!")]
+        [DataType(DataType.Text, ErrorMessage = "Неправильные данные!")]
         public string SenderAddress { get; set; }
 
         [Required(ErrorMessage = "Поле обязательно")]
         [DisplayName("Город получателя")]
-        [DataType(DataType.Text,ErrorMessage = "Неправильные данные!")]
+        [DataType(DataType.Text, ErrorMessage = "Неправильные данные!")]
         public string RecipientCity { get; set; }
 
         [Required(ErrorMessage = "Поле обязательно")]
         [DisplayName("Адрес получателя")]
-        [DataType(DataType.Text,ErrorMessage = "Неправильные данные!")]
+        [DataType(DataType.Text, ErrorMessage = "Неправильные данные!")]
         public string RecipientAddress { get; set; }
 
         [Required(ErrorMessage = "Поле обязательно")]
@@ -34,7 +34,7 @@
 
         [Required(ErrorMessage = "Поле обязательно")]
         [DisplayName("Дата забора груза")]
-        [DataType(DataType.DateTime,ErrorMessage = "Неправильные данные!")]
+        [DataType(DataType.DateTime, ErrorMessage = "Неправильные данные!")]
         public DateTime PickupDate { get; set; }
     }
 }
